requirements.txt
abstract-rendering==0.5.1
alabaster==0.7.3
alembic==0.8.3
argcomplete==0.8.9
astropy==1.0.3
Babel==1.3
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
bat-country==0.2
bcolz==0.9.0
beautifulsoup4==4.3.2
binstar==0.11.0
bitarray==0.8.1
blaze==0.8.0
blz==0.6.2
bokeh==0.9.0
boto==2.38.0
Bottleneck==1.0.0
cdecimal==2.3
certifi==14.5.14
cffi==1.1.0
clyent==0.3.4
colorama==0.3.3
conda==4.0.5
conda-build==1.14.1
conda-env==2.4.5
configobj==5.0.6
cryptography==0.9.1
cycler==0.10.0
Cython==0.24
cytoolz==0.7.3
dask==0.8.1
datashape==0.4.5
decorator==3.4.2
docopt==0.6.2
docutils==0.12
dominate==2.1.16
enum34==1.0.4
fastcache==1.0.2
Flask==0.10.1
Flask-Bootstrap==3.3.5.6
Flask-Migrate==1.6.0
Flask-Moment==0.5.1
Flask-Script==2.0.5
Flask-SocketIO==0.6.0
Flask-SQLAlchemy==2.0
Flask-WTF==0.12
funcsigs==0.4
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.7
grin==1.2.1
gunicorn==17.5
h5py==2.6.0
idna==2.0
imutils==0.2.2
ipaddress==1.0.7
ipykernel==4.1.1
ipython==4.1.2
ipython-genutils==0.1.0
ipywidgets==4.1.0
itsdangerous==0.24
jdcal==1.0
jedi==0.8.1
Jinja2==2.8
jsonschema==2.4.0
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.0.3
jupyter-core==4.0.6
leveldb==0.193
llvmlite==0.5.0
lmdb==0.89
lxml==3.4.4
Mako==1.0.2
MarkupSafe==0.23
matplotlib==1.4.3
mistune==0.7.1
mock==1.0.1
multipledispatch==0.4.7
nbconvert==4.0.0
nbformat==4.0.1
networkx==1.9.1
nltk==3.0.3
nose==1.3.7
notebook==4.0.6
numba==0.19.1
numexpr==2.5.1
numpy==1.11.1
odo==0.3.2
openpyxl==1.8.5
pandas==0.18.0
pandoc==1.0.0b2
path.py==0.0.0
patsy==0.3.0
PDF==1.0
pep8==1.6.2
pexpect==3.3
pickleshare==0.5
Pillow==2.9.0
ply==3.6
powerline-status==2.2
protobuf==2.6.1
psutil==2.2.1
ptyprocess==0.5
py==1.4.27
pyasn1==0.1.7
pycairo==1.10.0
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.19.5.1
pydot2==1.0.33
pyflakes==0.9.2
Pygments==2.0.2
pyOpenSSL==0.15.1
pyparsing==2.0.3
pytest==2.7.1
python-dateutil==2.4.2
python-editor==0.4
python-gflags==2.0
pytz==2015.4
PyYAML==3.11
pyzmq==14.7.0
qtconsole==4.1.1
redis==2.10.3
requests==2.9.1
rodeo==0.4.4
rope==0.9.4
runipy==0.1.3
scikit-image==0.11.3
scikit-learn==0.17.1
scipy==0.15.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.9.0
snowballstemmer==1.2.0
sockjs-tornado==1.0.1
Sphinx==1.3.1
sphinx-rtd-theme==0.1.7
spyder==2.3.6
SQLAlchemy==1.0.8
statsmodels==0.6.1
sympy==0.7.6
tables==3.2.0
terminado==0.5
Theano==0.7.0
toolz==0.7.2
tornado==4.3
traitlets==4.0.0
ujson==1.33
unicodecsv==0.9.4
virtualenv==13.0.1
Werkzeug==0.11.2
wheel==0.24.0
WTForms==2.0.2
xlrd==0.9.3
XlsxWriter==0.7.3
xlwt==1.0.0
